Acq/File Info extensions (#1)
Create a framework for extensible AcqType and FileType

Extension types are for adding new types of acquisition and files. There are done providing a class implementing AcqInfoBase or FileInfoBase which contains methods for determining if a file is of this type, and setting any required metadata. A concrete implementation can be found in alpenhorn.generic.