Allow nested acquisition directory and file structure (fixes #4)
This change removes the assumption that all acquisitions are directories of the Alpenhorn root and contain files without any nesting of directories. The key is moving the decision of what is the acquisition name out of `auto_import` and into `AcqType.detect`, by finding the first registered acquisition type that handles a closest possible ancestor directory of the file being imported. The rest of the changes follow in a straightforward manner.