Speed up test Docker image build by maximizing cacheable layers in Dockerfile
If the command is the same, and (for COPY/ADD) the underlying file hasn't
changed, Docker will reuse the cached old layer when rebuilding the image. So by
putting anything related to the actual Alpenhorn code to the very end of the
Docker file, most of the time everything else can be reused, drastically
speeding up the build.