Speed up testing acq sync by batching the running requests and then check
The original test case would issue a sync request, wait for 3 seconds, and then
verify that the files have been copied, and repeat this three times, for each of
the fixture acquisitions. Instead, we can issue syncs, wait for 3s, and then
check for all three acqs, saving 6s of the running time.