Use 'active' instead of 'mounted' in user commands (#43)
Alpenhorn does not perform the filesystem mounting of regular storage nodes, so
the client-facing terminology should reflect that. The "mounting" process is
really just activating an already mounted filesystem as an Alpenhorn storage
node.

(This does not apply to commands operating on transport disks, which do perform
the filesystem mounting, so we keep their existing names.)