Fix #9: running `import_files` from any subdirectory of Alpenhorn root
`client.import_files` has a number of built-in assumptions left over from the
old world of flat acquisitions, so it would not work properly when run from any
subdirectory of an arbitrarily nested structure (even subdirectories of an
acquisition).

This change implements the "natural" extension of the original `import_files`
design to Alpenhorn v2 directory structure. The starting directory can be inside
an acquisition or a parent directory of several acquisitions, but either way it
only considers for import files that are found by `os.walk` from that starting
point.

If `--acq` option is given, the walk is further restricted to only look at files
that would belong to those acquisitions.