Fix #118: `verify` called `alpenhorn_node_check` with wrong argument
`util.alpenhorn_node_check`'s argument is a `StorageNode`, not its name as
`client.verify` was passing. Also, it's now up to the caller to check that the
node is active, which was missing in `verify`.

Finally, the test for the `verify` method mocked `alpenhorn_node_check`, so this
bug was masked in testing. I removed the mock and test all code paths
realistically.