Add a command to list transport nodes
Since transport disks are a type of storage node, renaming and modifying them
are already available in the 'node' subcommands. This version of the `list`
command is tuned for transport disks in that it selects by 'storage_type=T' and
shows slightly different columns (mounted, but not group or type).